Fix some compiler warnings.
Use PRId64 macros to print 64-bit ints.
Unfortunately 64-bit Linux complains about using %lld for uint64_t.
Be more explicit with some bitwise order of operations.
Use UINT32_MAX for initializing 32-bit values.

+++ media/base/shell_buffer_factory.cc
@@ -284,7 +284,7 @@ ShellBufferFactory::~ShellBufferFactory() {
   if (allocs_.size() > 1) {
     DLOG(WARNING) << base::StringPrintf(
         "%d unfreed allocs on termination now pointing at invalid memory!",
-        allocs_.size() - 1);
+        static_cast<int>(allocs_.size()) - 1);
   }
   // and no outstanding array requests
   DCHECK_EQ(array_requested_size_, 0);
